pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentctl"
version = "0.1.0"
description = "A lightweight CLI for managing AI agents across providers. kubectl for AI agents."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Ander Turing", email = "anderturing@gmail.com" },
]
keywords = ["ai", "agents", "cli", "llm", "openai", "anthropic", "ollama"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "httpx>=0.25",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.18"]
all = ["openai>=1.0", "anthropic>=0.18"]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21", "ruff>=0.1"]

[project.scripts]
agentctl = "agentctl.cli:main"

[project.urls]
Homepage = "https://github.com/anderturing-debug/agentctl"
Repository = "https://github.com/anderturing-debug/agentctl"
Issues = "https://github.com/anderturing-debug/agentctl/issues"

[tool.ruff]
line-length = 100
target-version = "py310"
